Revert "feat: added optional code interpreter tool (#102)"
This reverts commit 397bb4a4aa74f550b5b01c5fcbacf1d99af9620b.
