+++ src/components/BotError.tsx
@@ -1,5 +1,4 @@
 import { AlertTriangle, ChevronRight } from 'lucide-react'
-import { CollapsibleSection } from './ui/collapsible-section'
 
 interface BotErrorProps {
   message: string
@@ -12,14 +11,16 @@ export function BotError({ message }: BotErrorProps) {
         <AlertTriangle className="h-5 w-5" />
       </div>
       <div className="flex flex-col space-y-1 items-start w-full sm:w-[85%] md:w-[75%] lg:w-[65%]">
-        <CollapsibleSection
-          title="Error"
-          icon={<AlertTriangle className="h-4 w-4" />}
-          variant="red"
-          open={true}
+        <details
+          open
+          className="rounded-2xl px-4 py-2 text-sm w-full group [&:not([open])]:h-8 [&:not([open])]:flex [&:not([open])]:items-center [&:not([open])]:py-0 bg-red-50 dark:bg-red-950"
         >
+          <summary className="font-medium mb-1 flex items-center gap-2 list-none [&::-webkit-details-marker]:hidden cursor-pointer group-[&:not([open])]:mb-0">
+            <ChevronRight className="h-4 w-4 transition-transform group-open:rotate-90" />
+            <span>Error</span>
+          </summary>
           <div>{message}</div>
-        </CollapsibleSection>
+        </details>
       </div>
     </div>
   )
