+++ src/components/CodeBlock.tsx
@@ -40,7 +40,7 @@ export function CodeBlock({
     <div className="relative group">
       <pre
         ref={preRef}
-        className={`overflow-auto max-h-64 md:max-h-80 lg:max-h-96 whitespace-pre-wrap break-words bg-gray-100 text-gray-900 dark:bg-gray-900 dark:text-gray-100 rounded-lg p-4 border border-gray-200 dark:border-gray-700 ${className || ''}`}
+        className={`overflow-x-auto whitespace-pre-wrap break-words ${className || ''}`}
         {...props}
       >
         {children}
@@ -49,7 +49,7 @@ export function CodeBlock({
         onClick={handleCopy}
         variant="secondary"
         size="sm"
-        className="absolute top-2 right-2 opacity-0 group-hover:opacity-100 transition-opacity duration-200 text-xs bg-white hover:bg-gray-50 text-gray-900 border-gray-300 dark:bg-gray-700 dark:hover:bg-gray-600 dark:text-gray-100 dark:border-gray-600"
+        className="absolute top-2 right-2 opacity-0 group-hover:opacity-100 transition-opacity duration-200 text-xs"
       >
         <Copy className="h-4 w-4" />
         <span className="sr-only">Copy code snippet</span>
