+++ src/components/ReasoningMessage.tsx
@@ -1,7 +1,6 @@
 import { Brain } from 'lucide-react'
 import { cn } from '../lib/utils'
 import { MarkdownContent } from './MarkdownContent'
-import { CollapsibleSection } from './ui/collapsible-section'
 
 type ReasoningMessageProps = {
   effort: string
@@ -35,13 +34,53 @@ export function ReasoningMessage({
       </div>
 
       <div className="flex flex-col space-y-1 items-start w-full sm:w-[85%] md:w-[75%] lg:w-[65%]">
-        <CollapsibleSection
-          title="Reasoning"
-          icon={<Brain className="h-4 w-4" />}
-          open={true}
-          isLoading={isLoading}
+        <details
+          open
+          className="rounded-2xl px-4 py-2 text-sm w-full bg-purple-50 text-purple-900 dark:bg-purple-950 dark:text-purple-100 group [&:not([open])]:h-8 [&:not([open])]:flex [&:not([open])]:items-center [&:not([open])]:py-0"
         >
-          <div className="text-xs space-y-1">
+          <summary className="font-medium mb-1 flex items-center gap-2 list-none [&::-webkit-details-marker]:hidden cursor-pointer group-[&:not([open])]:mb-0">
+            {/* Chevron icon for expand/collapse */}
+            <svg
+              className="h-4 w-4 transition-transform group-open:rotate-90 text-purple-600 dark:text-purple-300"
+              fill="none"
+              stroke="currentColor"
+              strokeWidth="2"
+              viewBox="0 0 24 24"
+              aria-hidden="true"
+            >
+              <path
+                strokeLinecap="round"
+                strokeLinejoin="round"
+                d="M9 5l7 7-7 7"
+              />
+            </svg>
+            Reasoning
+            {isLoading && (
+              <span className="ml-2 animate-spin">
+                <svg
+                  className="h-4 w-4"
+                  viewBox="0 0 24 24"
+                  fill="none"
+                  aria-label="Loading"
+                >
+                  <circle
+                    className="opacity-25"
+                    cx="12"
+                    cy="12"
+                    r="10"
+                    stroke="currentColor"
+                    strokeWidth="4"
+                  />
+                  <path
+                    className="opacity-75"
+                    fill="currentColor"
+                    d="M4 12a8 8 0 018-8v4a4 4 0 00-4 4H4z"
+                  />
+                </svg>
+              </span>
+            )}
+          </summary>
+          <div className="text-xs space-y-1 mt-1">
             {effort && <div>Effort: {effort}</div>}
             {summary && <MarkdownContent content={summary} />}
             {model && <div>Model: {model}</div>}
@@ -53,7 +92,7 @@ export function ReasoningMessage({
               {topP !== undefined && <div>Top P: {topP}</div>}
             </div>
           </div>
-        </CollapsibleSection>
+        </details>
       </div>
     </div>
   )
