+++ src/components/ToolCallMessage.tsx
@@ -1,6 +1,5 @@
 import { getStatusIcon } from '@/lib/toolStatus'
 import { Wrench, ChevronRight, ArrowRight } from 'lucide-react'
-import { CollapsibleSection } from './ui/collapsible-section'
 
 type ToolCallMessageProps<T = Record<string, unknown>> = {
   name: string
@@ -52,13 +51,6 @@ const getTitle = (name: string, args: ToolCallMessageProps['args']) => {
 }
 
 export function ToolCallMessage({ name, args }: ToolCallMessageProps) {
-  const summaryContent = (
-    <>
-      {getStatusIcon(args.status, args.error)}
-      <span className="sr-only">{getStatusText(args.status, args.error)}</span>
-    </>
-  )
-
   return (
     <div className="flex w-full max-w-full gap-2 py-2 animate-in fade-in justify-start">
       <div className="flex h-8 w-8 shrink-0 select-none items-center justify-center rounded-md bg-orange-100 text-orange-600 dark:bg-orange-900 dark:text-orange-300">
@@ -66,17 +58,20 @@ export function ToolCallMessage({ name, args }: ToolCallMessageProps) {
       </div>
 
       <div className="flex flex-col space-y-1 items-start w-full sm:w-[85%] md:w-[75%] lg:w-[65%]">
-        <CollapsibleSection
-          title={getTitle(name, args)}
-          icon={<Wrench className="h-4 w-4" />}
-          variant="orange"
-          additionalSummaryContent={summaryContent}
-        >
+        <details className="rounded-2xl px-4 py-2 text-sm w-full group [&:not([open])]:h-8 [&:not([open])]:flex [&:not([open])]:items-center [&:not([open])]:py-0 bg-orange-50 text-orange-900 dark:bg-orange-950 dark:text-orange-100">
+          <summary className="font-medium mb-1 flex items-center gap-2 list-none [&::-webkit-details-marker]:hidden cursor-pointer group-[&:not([open])]:mb-0">
+            <ChevronRight className="h-4 w-4 transition-transform group-open:rotate-90" />
+            {getTitle(name, args)}
+            {getStatusIcon(args.status, args.error)}
+            <span className="sr-only">
+              {getStatusText(args.status, args.error)}
+            </span>
+          </summary>
           {args.error && <span>Error: {args.error}</span>}
           <pre className="text-xs overflow-x-auto whitespace-pre-wrap break-words break-all">
             {JSON.stringify(args, null, 2)}
           </pre>
-        </CollapsibleSection>
+        </details>
       </div>
     </div>
   )
