+++ src/routeTree.gen.ts
@@ -13,7 +13,6 @@ import { createServerRootRoute } from '@tanstack/react-start/server'
 import { Route as rootRouteImport } from './routes/__root'
 import { Route as IndexRouteImport } from './routes/index'
 import { ServerRoute as ApiModelsServerRouteImport } from './routes/api/models'
-import { ServerRoute as ApiContainerFileServerRouteImport } from './routes/api/container-file'
 import { ServerRoute as ApiChatServerRouteImport } from './routes/api/chat'
 
 const rootServerRouteImport = createServerRootRoute()
@@ -26,11 +25,6 @@ const IndexRoute = IndexRouteImport.update({
 const ApiModelsServerRoute = ApiModelsServerRouteImport.update({
   id: '/api/models',
   path: '/api/models',
-  getParentRoute: () => rootServerRouteImport,
-} as any)
-const ApiContainerFileServerRoute = ApiContainerFileServerRouteImport.update({
-  id: '/api/container-file',
-  path: '/api/container-file',
   getParentRoute: () => rootServerRouteImport,
 } as any)
 const ApiChatServerRoute = ApiChatServerRouteImport.update({
@@ -62,31 +56,27 @@ export interface RootRouteChildren {
 }
 export interface FileServerRoutesByFullPath {
   '/api/chat': typeof ApiChatServerRoute
-  '/api/container-file': typeof ApiContainerFileServerRoute
   '/api/models': typeof ApiModelsServerRoute
 }
 export interface FileServerRoutesByTo {
   '/api/chat': typeof ApiChatServerRoute
-  '/api/container-file': typeof ApiContainerFileServerRoute
   '/api/models': typeof ApiModelsServerRoute
 }
 export interface FileServerRoutesById {
   __root__: typeof rootServerRouteImport
   '/api/chat': typeof ApiChatServerRoute
-  '/api/container-file': typeof ApiContainerFileServerRoute
   '/api/models': typeof ApiModelsServerRoute
 }
 export interface FileServerRouteTypes {
   fileServerRoutesByFullPath: FileServerRoutesByFullPath
-  fullPaths: '/api/chat' | '/api/container-file' | '/api/models'
+  fullPaths: '/api/chat' | '/api/models'
   fileServerRoutesByTo: FileServerRoutesByTo
-  to: '/api/chat' | '/api/container-file' | '/api/models'
-  id: '__root__' | '/api/chat' | '/api/container-file' | '/api/models'
+  to: '/api/chat' | '/api/models'
+  id: '__root__' | '/api/chat' | '/api/models'
   fileServerRoutesById: FileServerRoutesById
 }
 export interface RootServerRouteChildren {
   ApiChatServerRoute: typeof ApiChatServerRoute
-  ApiContainerFileServerRoute: typeof ApiContainerFileServerRoute
   ApiModelsServerRoute: typeof ApiModelsServerRoute
 }
 
@@ -110,13 +100,6 @@ declare module '@tanstack/react-start/server' {
       preLoaderRoute: typeof ApiModelsServerRouteImport
       parentRoute: typeof rootServerRouteImport
     }
-    '/api/container-file': {
-      id: '/api/container-file'
-      path: '/api/container-file'
-      fullPath: '/api/container-file'
-      preLoaderRoute: typeof ApiContainerFileServerRouteImport
-      parentRoute: typeof rootServerRouteImport
-    }
     '/api/chat': {
       id: '/api/chat'
       path: '/api/chat'
@@ -135,7 +118,6 @@ export const routeTree = rootRouteImport
   ._addFileTypes<FileRouteTypes>()
 const rootServerRouteChildren: RootServerRouteChildren = {
   ApiChatServerRoute: ApiChatServerRoute,
-  ApiContainerFileServerRoute: ApiContainerFileServerRoute,
   ApiModelsServerRoute: ApiModelsServerRoute,
 }
 export const serverRouteTree = rootServerRouteImport
